build: include more files in "make dist" tarball
Include all the documentation, qa, and CMakeLists.txt files in the "make
dist" tarball.

Kraken and newer releases will already do this. This change is
restricted to releases that still use autotools.

The purpose of this change is to make it easier to apply downstream
patches on an unpacked upstream tarball, because often those patches are
cherry-picks of commits that touch files under doc/, qa/, or
CMakeLists.txt.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>