osd_types,PG: force the map used in _scrub to be sorted correctly
Previously, ScrubMap::objects was always sorted bitwise (nibblewise
before the comparator change was made.  It didn't matter because we
always scrubbed whole hash values.  Now, we need the objects in the
objectstore ordering because we may be missing objects at the tail of
the scanned range and need them to show up at the tail of the
ScrubMap::objects mapping.  We don't need to do anything else to handle
the upgrade process since the actual objects *in* the map were
determined by the objectstore ordering.

Signed-off-by: Samuel Just <sjust@redhat.com>