librbd / ObjectMap: multiview ObjectMap boilerplates
Maintain an multi-view in-memory version on ObjectMap that's
used to batch ObjectMap update requests. View (ObjectMapView
class) consist of a master copy of the ObjectMap and a set of
(one or more) tracking states.

Batch updates to the ObjectMap are applied on a particular view
(or levels) after satisfying update requirements (as usual) for
a given object umber. Object state from the latest view is used
when querying the ObjectMap. During batch updations, the master
copy of the ObjectMap remains untouched as all updates are
batched (or "tracked") in their respective views(s). In-memory
view(s) are synchronized to disk by first applying each view
to the master copy ObjectMap (so as to bring the ObjectMap in
sync with the current state) followed by updating it on-disk
sequentially. Flushing batched updates is controlled by the
caller and is not done internally now (however that should be
doable without much fuss).

Signed-off-by: Venky Shankar <vshankar@redhat.com>