mon: assume old mons that didn't specific kv type are leveldb
This assumes that if the mon does not explicitly specify
the kv type that it is leveldb.  No prior version of
Ceph has had non-experimental rocksdb, so this is
relatively safe.  It's also necessary because the
default is now 'rocksdb' and we shouldn't assume those
old mons are rocksdb.

This will break for users to explicitly specified
rocksdb for the mon despite it being experimental.

Signed-off-by: Sage Weil <sage@redhat.com>