librbd / ObjectMap: assert on interfering batch updates
When batch updations are in progress, non-batched updates
can result in incorrect values to be persisted (or worse,
get overwritten) in the on-disk ObjectMap. Such updates
are caught (if the non-batched update falls in the
operating "batched" range) and asserted for now. Further,
this scenario can be dealt with by flushing the batched
requests followed by applying the non-batched updated.

Signed-off-by: Venky Shankar <vshankar@redhat.com>