orchestra.connect: do not override user/host from ~/.ssh/config
I don't know why this was ever the case, but ignoring user@host
from the targets list just seems like a bad idea, and causes
problems, because tasks/ansible/etc. assume that the user to use
is ubuntu.  If that's not what .ssh/config says, there's just
going to be trouble...but there's no reason to pay attention
to .ssh/config there; if you really want to play tricks, edit
your target list accordingly.

Also, the original code was apparently not prepared for the
key_filename to be a list, which it apparently always is,
at least now; handle that too.

Fixes: http://tracker.ceph.com/issues/17125
Signed-off-by: Dan Mick <dan.mick@redhat.com>