mon: send updated monmap to its subscribers
prior to this change, we sends monmap when serving the subscription
requests, but the updates are not sent to the subscribers anymore.
so we need to send latest monmap to subscribes and update the
subscription status accordingly when the monmap is updated.

http://tracker.ceph.com/issues/17558
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit c2b348b12c47ba505f72228ae0ea9b1a1b674719)

Conflicts:
	src/mon/Monitor.cc (trivial resolution)
	src/mon/MonmapMonitor.cc (in master, check_subs() and check_sub() were
            refactored using C++11isms; for the hammer backport, we revert to the
            original implementation, taking into account that the functions now
            live in the MonmapMonitor class)