client: allow UserPerm constructor to populate gid list
Add args for the gids_count and gids list, and give them default
values so that callers can populate it correctly. We'll need this
for ganesha so it can populate the UserPerm from a RPC AUTH object.

Note that the gidlist pointer must be valid for the lifetime of
the created object!

Signed-off-by: Jeff Layton <jlayton@redhat.com>