client: change args to ceph_readdirplus_r
Make ceph_readdirplus_r take a ceph_statx, a want and a flags parm. With
this, we can allow applications to express an interest in subset of the
attributes, and can allow for a "lazy" readdirplus.

Drop the stmask. It ends up returning the caps that the client holds on
the inode. That's not well defined, and we can now express that in a
better way via the stx_mask, which applications can use to tell which
fields in the ceph_statx are actually valid.

For now, the want mask is ignored. I don't see a way to ask for a set
of caps in a ceph readdir request on the wire. Maybe we could add that?

Signed-off-by: Jeff Layton <jlayton@redhat.com>