client: pass want and flags to readdir_r_cb
...so we can ensure that we have the necessary caps when filling out
the ceph_statx for each dentry's inode. In order to only do this when
completely necessary, we have want default to 0 and the flags default
to AT_NO_ATTR_SYNC. The only codepath where we pass in a non-default
set of args there is ceph_readdirplus_r as it's the only codepath that
cares about fields in the ceph_statx that aren't immutable.

For now, since we have no support for requesting caps during a readdir
call, we simply issue getattrs prior to calling fill_statx. If we
already have the necessary caps, or are doing a lazy statx then this
becomes a no-op.

Note too that I _think_ the MDS will recall caps on the entries when
satisfying a readdir, so we avoid calling getattr when we're populating
the ceph_statx out of a just-acquired readdir response.

Signed-off-by: Jeff Layton <jlayton@redhat.com>