rgw: implement buffered HTTP header composition in front-ends.
Without the patch front-ends issueed a lot of small IO operations
leading to increased overhead on syscalls and to the fragmentation
of an HTTP message across multiple TCP segments. The later was
occuring when the Nagle's algorithm hadn't been able to form
a single TCP segment (usually when running on extremely fast
network interfaces like loopback).

Signed-off-by: Radoslaw Zarzynski <rzarzynski@mirantis.com>