mon: Do not allow pools to be deleted by default
This prevents accidental pool removals on cluster which may contain
very valuable data.

Users can still revert this setting and allow a pool to be removed,
but that would be a deliberate action of the admin.

It can either be enabled by setting it in the [mon] section of the
ceph.conf configuration file or on runtime using ceph tell:

Signed-off-by: Wido den Hollander <wido@42on.com>