Make PG::queue_op lockless for common case
WIP patch, more extensive tests needed..

PG::queue_op won't queue a request if the epoch is newer
than the one in the request. This patch changes that check
to be done in a lockless manner.

(1) obtaining osdmap using an RCU managed pointer
(2) manage the osdmap's epoch using an atomic counter
(3) maintain request ordering for the session

Signed-off-by: Dan Lambright <dlambrig@gmail.com>