crushtool: Allow reweighting to desired multidraw distribution
CRUSH correctly performs repeated weighted draws. I had thought it was
improperly weighting items in subsequent draws, and patch seemed to
work, but it was a fixed-point error giving a false success on the
first examples I tried. Testing it demonstrated that it wasn't
working, and fixing the bug brought the old distribution back.
When one compares its output to the calculated probability, CRUSH
approximates it well.

The distribution CRUSH yields can be surprising to users. They may
expect that if one item has a weight of N and another a weight of 1,
there should be N times more objects in the first device than the second.

A device weighted N is N times more likely in the first draw.
Subsequent draws add to that. We cannot adjust the weights in
subsequent draws: the ratio would approach what we expect as the
adjusted weight goes to infinity. This is terrible.

We can iteratively find single-draw weights to approximate the desired
multidraw distribution. This adjustment depends on the replication
count, which is only known at run-time. We should not do this
adjustment at run-time. It is time consuming. Worse, changing the
distribution with replication count throws out the monotonicity
guarantee our bucket algorithms work hard to provide.

Additionally, the current behavior isn't actually wrong, merely
surprising. We don't want to change it for all users.

The crushtool utility now has the --tweak-bucket-weights option.
Given a bucket with weights a, b, c,...z where z is the minimum and a
draw count r, it adjusts the bucket's weights so that for draws of r
items, the a-weighted item appears in approximately a/z times as many
r-draws and so forth.

Distributions are simply not possible. Given a two item draw from
three items, there is simply no way for one item to show up in ten
times as many draws as the other two items.  In these cases, crushtool
will exhaust the allowed number of tries (specified with
--tweak-tries) and stop. (I've ruled out the most egregiously
impossible case. The general case would take lots of compute-time and
likely not be worth it.)

It is not worth using this feature on large (relative to the
replication count) buckets with a few items departing from an
otherwise uniform weight; they will approximate the desired
distribution on their own.

This should be a special, separate operation. It specializes CRUSH
maps for a given replication count in a way that did not happen
before.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>