common,test: g_ceph_context->put() upon return
prior to this change, global_init() could create a new CephContext
and set assign it to g_ceph_context. it's our responsibilty to release
the CephContext explicitly using cct->put() before the application
quites. but sometimes, we fail to do so. in this change, global_init()
will return an intrusive_ptr<CephContext>, which calls
`g_ceph_context->put()` in its dtor. this ensures that the CephContext
is always destroyed before the application quites. and hence flushes the
log before _log_exp_length is destroyed.

there are two cases where global_pre_init() is called directly.
- ceph_conf.cc: g_ceph_context->put() will be called by an intrusive_ptr<>
  deleter.
- rgw_main.cc: global_init() is called later on on the success code
  path, so it will be taken care of.

Fixes: http://tracker.ceph.com/issues/17762
Signed-off-by: Kefu Chai <kchai@redhat.com>