ceph-disk: Adding cluster name support for dmcrypt
Prior to this commit we were not able to configure an OSD using dmcrypt
on a cluster with a different name than 'ceph'.

Wwe have to propagate the cluster name (args.cluster) to the get_dmcrypt_key()
fonction, which imply changing the prototype of it by adding a "cluster" argument.

This changes imply that dmcrypt_map(), mount_activate(),  main_activate_space()
functions prototypes are changing to get the cluster name for their respective
callers. It's usually a way just to make args.cluster available for the
functions they call. Note that we introduce the "cluster" argument of
mount_activate() _before_ the last argument (reactivate) as this one have a
default value : as we don't have any for "cluster", that would make a python
exception.

This patch has been tested with the following commands :
- deactivate
- destroy
- prepare

Fixes: http://tracker.ceph.com/issues/17821
Signed-off-by: Sébastien Han <seb@redhat.com>
Signed-off-by: Erwan Velu <erwan@redhat.com>