common: Add throttle_get_started perf counter
This change adds 'get_started' performance counter which
is increased at the beginning of Throttle::get() call,
before any potential wait is started. It's purpose is
to measure in realtime the number of jobs waiting for the
throttle (get_started - get) and get the characteristic of
get cals before throttler does it's job.

Signed-off-by: Bartłomiej Święcki <bartlomiej.swiecki@corp.ovh.com>