ceph_test_objectstore: fix Rename test
We were doing a no-op setattr here, even though the intenral
ENOENT used to be ignored, because it previously triggered a
bug in bluestore.  We're no longer ignoring that internal
error, though, as of 6ecb44f0c6216f3fb29b384a2d10c7d33bf1d19c,
so remove this part of the test.

The original bluestore bug we used to trigger was fixed by
0210f1fde6f8e17affdd0c49055a71a01f109447.

Signed-off-by: Sage Weil <sage@redhat.com>