ceph_test_objectstore: fix Rename test
We were doing a no-op setattr here, even though the intenral
ENOENT used to be ignored, because it previously triggered a
bug in bluestore.  We're no longer ignoring that internal
error, though, as of 8cc9f314c4975caa357727392abe13fbca6bb7c9,
so remove this part of the test.

The original bluestore bug we used to trigger was fixed by
8f9d4c03eba876a7d0fe19291d87dedaf906f324.

Signed-off-by: Sage Weil <sage@redhat.com>