os/bluestore/BlueFS: pre-extend file size for WAL (.log) files
When rocksdb has log recycling on (this is required!), it will
do robust checksums on log records and change playback behavior
to tolerate trailing garbage in the log file.  This normally
allows it to overwrite previous log files, but it can also let
us overwrite arbitrary garbage on the device.

If we allocate some new space for a .log file (already indicated
by the WRITER_WAL hint), extend the size immediately so that each
subsequent append doesn't have to (unless/until we do another
allocation).

This is safe as long as rocksdb recycling is enabled (which it is
by default).

This is faster because we don't have to flush the bluefs log on
every log append during the period after startup before rocksdb
starts recycling log files.

Signed-off-by: Sage Weil <sage@redhat.com>