osd: remove down osd from actingbackfill when remove_down_peer_info
When advacing map, and the acting and up set don't change in the new map,
the PG won't go through peering. The down osd are removed from the peer_info.
If there are async recovery targets in actingbackfill, which is also down,
they should be removed from actingbackfill as well.

Signed-off-by: Zhiqiang Wang <zhiqiang@xsky.com>