safe_io: always overwrite existing file
even if the existing file already has correct content. this helps to
re-create the SELinux context of the file being written. if there is a
file already exists with the same content but with a different SELinux
context, prior to this change, the file will keep unchanged. and its
SELinux context is preserved. so SELinux denial will be reported if the
application tries to read/write that file later.

Fixes: http://tracker.ceph.com/issues/16800
Signed-off-by: Kefu Chai <kchai@redhat.com>