client: fix the cross-quota rename boundary check conditions
We were previously rejecting a rename if either of the involved directories
was a quota root, even if the other directory was part of the same quota
"tree". What we really want to do is identify the correct quota root
(whether local or ancestral) for each directory and compare them. So
now we do.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 8e8892aa46accb519faa4bb9fecf66618f1b11b2)