include/Context: add sync_complete() hook
Sometimes we are able to complete a context synchronously, within the same
callchain of the caller who queued it.  In this case the locking rules are
usually a bit different for the caller.  Add a generic Context hook that
allows this.

If a sync-capable Context implements sync_complete, it can do whatever
it needs to do for this particular event.  If it is not implemented, the
generic implementation will return false, and the caller can use the
normal complete() as it normally would have (presumably by calling it
asynchronously, e.g., in a Finisher).

If sync_complete() is implemented, it *must* return true.

Signed-off-by: Sage Weil <sage@redhat.com>