osd: trim pglog when pg is recovered
In the current code, primary initiates pglog trim or peers inform
primary to trim when finishing recovery of the last object on it.
However, the pg is still in degraded/recovering/backfilling state
at that time and then the max log entries are kept. If there are
no IOs on this pg, it will hold up to the max number of log entries
in the memory, even if it's totally recovered.

Signed-off-by: Zhiqiang Wang <zhiqiang@xsky.com>