osd: fix fiemap inconsistency during scrub
Currently, because of some reasons, Although the content in object is consistent,
the allocation blocks on filesystem may be inconsistent,
which lead to different filesystem space usage for a same object.

Different space usage for same objects may cause huge inbalance based on current crush alg.
Therefore, we have enough confidence to consider this case as an inconsistency and report
a scrub error and repair it.

Signed-off-by: yaoning <yaoning@unitedstack.com>