denc: Enforce boundedness at compile time
Givne a type T, our desire is that if denc_traits<T>::bounded is true
that the function denc_traits<T>::bound_encode() execute while depending
on no particulars of the object passed to it.

This was why some bound_encode() functions enabled by a true bound trait
would pass a value of *(const T*)nullptr, causing a crash at run time if
the referenced were ever accessed in the called function. This is
undefined behavior and there is some concern it might be triggered
spuriously at low optimization levels.

This is a requirement that can be enforced at compile-time. To do so I
have set up a template type, bounded_t<T>, that can participate in
argument deduction and which can swallow references of type T passed to
it, functioning as a 'trap door' that allows the 'bounded' functions to
be easily called from non-bounded ones but not vice-versa.

This works quite well. The main downside is that the requirement that
bound_encode() for bounded types not actually receive a reference to the
object to be encoded means that a single template function cannot serve
as bound_encode, encode, and decode in one. A separate implementation of
bound_encode must be provided. This follows from the encode and decode
being written in terms of members of the class to be encoded and
decoded.

This seems to be unavoidable in the general case, unfortunately.

But I bet I could come up with a template based solution using members
to pointer variables for the specific case of feature-conditioned
serializing a sequence of class members.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>